Moving inner notif content to src section
