+++ src/angular/notifications/notification-inner-content-example.component.ts
