+++ src/angular/notifications/notification.module.ts
@@ -6,8 +6,6 @@ import {CommonModule} from "@angular/common";
 import {NotificationComponent} from "./notification/notification.component";
 import {NotificationContainerComponent} from "./container/notifcontainer.component";
 import {NotificationsService} from "./services/notifications.service";
-// import {NotificationContentComponent} from "./content/content.component";
-import {InnerNotifContent} from "../../../stories/ng2-component-lab/components/notification-inner-content-example.component";
 import {CreateDynamicComponentService} from "../utils/create-dynamic-component.service";
