+++ stories/ng2-component-lab/components/notifications-example.component.ts
@@ -4,7 +4,7 @@
 import {Component, Input, ViewChild} from "@angular/core";
 import {NotificationsService} from "../../../src/angular/notifications/services/notifications.service";
 import {NotificationSettings} from "../../../src/angular/notifications/utilities/notification.config";
-import {InnerNotifContent} from "./notification-inner-content-example.component";
+import {InnerNotifContent} from "../../../src/angular/notifications/notification-inner-content-example.component";
 
 @Component({
     selector: "notifications-example",
